rename conda.tlz to conda.common.iterators
