+++ conda/common/iterators.py
