+++ conda/common/pkg_formats/python.py
@@ -21,7 +21,7 @@ try:
 except ImportError:
     from conda._vendor.toolz.itertoolz import concat, concatv
 
-from conda.tlz import groupby_to_dict as groupby
+from conda.common.iterators import groupby_to_dict as groupby
 
 from ... import CondaError
 from ..compat import odict, open
