+++ conda/core/subdir_data.py
@@ -22,7 +22,7 @@ try:
 except ImportError:
     from conda._vendor.toolz.itertoolz import concat, take
 
-from conda.tlz import groupby_to_dict as groupby
+from conda.common.iterators import groupby_to_dict as groupby
 
 from .. import CondaError
 from ..auxlib.ish import dals
