+++ conda/models/match_spec.py
@@ -15,7 +15,7 @@ try:
 except ImportError:
     from conda._vendor.toolz.itertoolz import concat, concatv
 
-from conda.tlz import groupby_to_dict as groupby
+from conda.common.iterators import groupby_to_dict as groupby
 
 from .channel import Channel
 from .version import BuildNumberMatch, VersionSpec
